+++ ElStock_Spring/src/main/java/com/elstock/controller/PersonTestController.java
@@ -1,0 +1,18 @@
+package com.elstock.controller;
+
+import com.elstock.entity.Person;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class PersonTestController {
+    @GetMapping(value = "/person")
+    public Person test(){
+        Person bean = new Person();
+        bean.setId("kim9");
+        bean.setName("김구");
+        bean.setAddress("마포구 공덕동");
+        bean.setSalary(10000);
+        return bean ;
+    }
+}
