+++ ElStock_Spring/src/main/java/com/elstock/entity/Person.java
@@ -1,0 +1,15 @@
+package com.elstock.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+@Getter
+@Setter
+@ToString
+public class Person {
+    private String id;
+    private String name;
+    private String address;
+    private Integer salary;
+}
