+++ ElStock_Spring/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+# 설정 포트 번호로써 기본은 값은 8080입니다.
+server.port=8989
+# mysql database 설정
+# shopping은 데이터 베이스 이름) create database shopping;으로 생성 가능
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/shopping?serverTimezone=UTC&characterEncoding=UTF-8
+spring.datasource.username=root
+spring.datasource.password=mysql
