--- src/main/java/com/elstock/entity/Employee.java
@@ -1,14 +1,0 @@
-package com.elstock.entity;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class Employee {
-    private String id ;
-    private String name ;
-    private Integer age ;
-}
