--- src/main/java/com/elstock/examples/ProductStatus.java
@@ -1,8 +1,0 @@
-package com.elstock.examples;
-
-// enum은 열거형 데이터를 표현할 때 사용하는 기법
-// 기호 상수(symbolic constant)
-public enum ProductStatus {
-    // 판매중, 품절
-    SELL, SOLD_OUT
-}
